Add waypoint system
Added a waypoint system to the game.

Waypoints are set in the main manager using the two arrays. It is turned into a dictionary that can be accessed by anything.

The currentWaypoint variable determines where the player will spawn next. It should be set before changing scenes.

The scene changer now allows you to specify the name of a waypoint to warp to it. These are all set up for every warp in the game now.